src/requirements.txt
falcon
PyJWT
PIL
pypdf2
pyvips
cx_Oracle